Moved restart function to class
