+++ lib/Ball.js
@@ -1,7 +1,8 @@
-  class Ball {
+const GamePiece = require('../lib/Game-Piece.js');
+
+class Ball extends GamePiece{
   constructor (x, y, radius, dx, dy) {
-    this.x = x;
-    this.y = y;
+    super(x,y);
     this.radius = radius;
     this.dx = dx;
     this.dy = dy;
