+++ lib/Block.js
@@ -1,9 +1,8 @@
-const Ball = require('../lib/Ball.js');
+const GamePiece = require('../lib/Game-Piece.js');
 
-class Block{
+class Block extends GamePiece{
   constructor(x, y) {
-    this.x = x;
-    this.y = y;
+    super(x,y);
     this.w = 85; 
     this.h = 30;
     this.color = '#73A557';
