+++ lib/Game.js
@@ -89,7 +89,7 @@ class Game {
     return this
   }
  
-   createArrayOfBlocks() {
+  createArrayOfBlocks() {
     if (this.level === 1) {
       for (let column = 0; column < this.blockColumnTotal; column++) {
         for (let row = 0; row < 3; row++) {
@@ -125,6 +125,20 @@ class Game {
     return this.blocks 
     }
     return this
+  }
+  
+  levelUpWindow() {
+    let endLevelWindow = document.querySelector('.hide-level-up');
+    let endUserScore = document.querySelector('.end-level-score');
+
+    endUserScore.innerText = this.userScore;
+    endLevelWindow.classList.remove('hide-level-up');
+    endLevelWindow.classList.add('level-up');
+  }
+
+  restartGame(parentSection) {
+    parentSection.classList.remove('level-up');
+    parentSection.classList.add('hide-level-up')
   }
 }
