+++ lib/Paddle.js
@@ -1,7 +1,8 @@
-class Paddle {
+const GamePiece = require('../lib/Game-Piece.js');
+
+class Paddle extends GamePiece{
   constructor( x, y, w, h) {
-    this.x = x;
-    this.y = y;
+    super(x,y);
     this.w = w;
     this.h = h;
   }
