+++ lib/index.js
@@ -3,8 +3,8 @@ const Ball = require('../lib/Ball.js');
 const Block = require('../lib/Block.js');
 const Game = require('../lib/Game.js');
 
-const paddle = new Paddle (400, 700, 250, 10);
-const ball = new Ball (300, 370, 20, 7, 7);
+const paddle = new Paddle (400, 700, 1000, 10);
+const ball = new Ball (300, 370, 20, 20, 20);
 const block =  new Block(...arguments);
 const game = new Game(...arguments);
 
@@ -105,13 +105,7 @@ function runLevelUp() {
 }
 
 function levelUpWindow() {
-  let endLevelWindow = document.querySelector('.hide-level-up');
-  let endUserScore = document.querySelector('.end-level-score');
-
-  endUserScore.innerText = game.userScore;
-  endLevelWindow.classList.remove('hide-level-up');
-  endLevelWindow.classList.add('level-up');
-
+  game.levelUpWindow();
   let restartLevelBtn = document.querySelector('.restart-level-button');
 
   restartLevelBtn.addEventListener('click', restartGame);
@@ -121,8 +115,7 @@ function levelUpWindow() {
 function restartGame() {
   let parentSection = this.closest('section');
 
-  parentSection.classList.remove('level-up');
-  parentSection.classList.add('hide-level-up')
+  game.restartGame(parentSection)
   game.pauseGame = false;
   gameLoop(); 
 }
